test: move hub download http proof into jest suite
